+++ cocos/core/3d/misc/ppm.ts
@@ -1,0 +1,7 @@
+
+/**
+ * save a color buffer to a PPM file
+ */
+export function toPPM (buffer: Uint8Array, w: number, h: number) {
+    return `P3 ${w} ${h} 255\n${buffer.filter((e, i) => i % 4 < 3).toString()}\n`;
+}
